+++ CHANGELOG.md
@@ -50,6 +50,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add redox theme
 * Add `Brush` handling gradients (respecting angles, directions, color codes)
 * Add fluent design inspired theme (dark and light)
+* Add message example
+* Remove widgets example
 
 ### 0.3.1-alpha3